requirements.txt
djangorestframework==3.11.2
Django==3.1.8
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
psycopg2-binary==2.8.4
dj-database-url==0.5.0
gunicorn==20.1.0
whitenoise==5.0.1
PyYAML==5.4
PyMySQL==1.0.2
django-cors-headers==3.10.0